feat: add telegram app
This app contains base classes to be used in this
project. They subclass the base classes of
django_telegram_app, but with project-specific fields
and behaviour.

The settings have also been update to include telegram
in INSTALLED_APPS, use a swapped telegram_settings model
and hide the ModelAdmin for telegram_settings model as we
use the inlined one on the user